requirements/base.txt
# Bleeding edge Django
django>=1.9

# Configuration
django-environ>=0.4.0
django-secure>=1.0.1
whitenoise>=2.0.6


# Forms
django-braces>=1.8.1
django-crispy-forms>=1.5.2
django-floppyforms>=1.6.0

# Models
django-model-utils>=2.4

# Images
Pillow>=3.0.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth>=0.24.1

# For the persistence stores
psycopg2>=2.6.1

# Unicode slugification
unicode-slugify>=0.1.3
django-autoslug>=1.9.3

# Time zones support
pytz>=2015.7

# Redis support
django-redis>=4.2.0
redis>=2.10.0



# Your custom requirements go here
djangorestframework>=3.3.1
django-guardian>=1.3
django-location-field
django-nested-admin
